+++ CHANGELOG.md
@@ -20,6 +20,12 @@
 
   - Previously leads to incorrect result space computation, and may lead to time slots missing despite meeting criteria
 
+## Timedesc 0.9.0
+
+- Moved sexp code into `timedesc-sexp` to further reduce core dependencies
+
+- Replaced use of mparser with angstrom
+
 ## Timedesc 0.8.0
 
 - Significantly reduced number of dependencies, and moved JS, JSON code into separate packages